refactor : 도메인 객체들을 하위 패키지로 분류
