+++ src/main/java/domain/player/GameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 public enum GameResult {
     WIN, LOSE, DRAW;
